requirements.txt
Adafruit-Blinka==8.67.0
adafruit-circuitpython-busdevice==5.2.14
adafruit-circuitpython-connectionmanager==3.1.6
adafruit-circuitpython-dht==4.0.10
adafruit-circuitpython-requests==4.1.15
adafruit-circuitpython-typing==1.12.3
Adafruit-PlatformDetect==3.84.1
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorzero==2.0
dataclasses==0.6
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gpiozero==2.0.1
httplib2==0.31.0
idna==3.10
oauthlib==3.3.1
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyftdi==0.57.1
pyparsing==3.2.5
pyserial==3.5
python-dotenv==1.1.1
pyusb==1.3.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
pytest>=7.0.0
